pyproject.toml
[project]
name = "it-learns-mazes-environment"
version = "0.0.1"
dependencies = [
  "gymnasium>=1.1.1",
  "numpy>=2.2.6",
  "pygame-ce>=2.5.3",
  "maturin>=1.8.62",
  "pytest>=8.3.5",
  "numpy>=2.2.6",
]
